fastapi-solution/postgres_to_es/requirements.txt
flake8==4.0
python-dotenv==0.20.0
psycopg2==2.9.3
elasticsearch==8.3.3
pytest-cov==3.0.0
isort==5.10.1
pydantic==1.9.1
